+++ CHANGELOG.md
@@ -39,7 +39,7 @@
 * Browse contributors | Second pane header doesn't update when user return to "Browse inventory" pane via the web-browser "Back" button. Fixes UIIN-2181.
 * Define new route for Inventory "Browse" page. Refs UIIN-2193.
 * Relabel 'Bound-with titles' accordion on item view. Fixes UIIN-2196.
-* Link from bound-with table HRID column to instances. Fixes UIIN-2195.
+* Link from bound-with table's instance HRID and holdings record HRID columns to the respective records. Fixes UIIN-2195.
 * Browse contributors | Show an indicator that contributor value is controlled/linked to an authority record. Refs UIIN-2179.
 * View Instance record | Contributors accordion | Show authorized indicator for each contributor. Refs UIIN-2180.
 * Move Inventory `Browse` logic into separate page: filters and result. Refs UIIN-2194.